flac: fix seeking precision

+++ plugins/flac/flac.c
@@ -583,7 +583,7 @@ cflac_seek_sample (DB_fileinfo_t *_info, int sample) {
 
 static int
 cflac_seek (DB_fileinfo_t *_info, float time) {
-    return cflac_seek_sample64 (_info, time * _info->fmt.samplerate);
+    return cflac_seek_sample64 (_info, (int64_t)((double)time * (int64_t)_info->fmt.samplerate));
 }
 
 static FLAC__StreamDecoderWriteStatus
